added modelstate.isvalid check to controller

+++ src/ZenithWebSite/Controllers/EventsApiController.cs
@@ -44,30 +44,45 @@ namespace ZenithWebSite.Controllers
 
         // POST api/values
         [HttpPost]
-        public void Post([FromBody]Event value)
+        public async Task<IActionResult> Post([FromBody]Event value)
         {
-            _context.Events.Add(value);
-            _context.SaveChanges();
+            if (ModelState.IsValid)
+            {
+                _context.Events.Add(value);
+                await _context.SaveChangesAsync();
+                return Ok();
+            }
+            return BadRequest();
         }
 
         // PUT api/values/5
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody]Event value)
+        public async Task<IActionResult> Put(int id, [FromBody]Event value)
         {
-            _context.Events.Update(value);
-            _context.SaveChanges();
+            if (ModelState.IsValid)
+            {
+                _context.Events.Update(value);
+                await _context.SaveChangesAsync();
+                return Ok();
+            }
+            return BadRequest();
         }
 
         // DELETE api/values/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public async Task<IActionResult> Delete(int id)
         {
-            var value = _context.Events.FirstOrDefault(t => t.EventId == id);
-            if (value != null)
+            if (ModelState.IsValid)
             {
-                _context.Events.Remove(value);
-                _context.SaveChanges();
+                var value = _context.Events.FirstOrDefault(t => t.EventId == id);
+                if (value != null)
+                {
+                    _context.Events.Remove(value);
+                    _context.SaveChanges();
+                }
+                return Ok();
             }
+            return BadRequest();
         }
     }
 }
